aded wrgb to mlx_util and changed int to size_t

+++ srcs/mlx_util/put_pixel.c
@@ -6,13 +6,13 @@
 /*   By: gcrisp <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/21 15:08:34 by gcrisp            #+#    #+#             */
-/*   Updated: 2025/02/24 14:42:33 by gcrisp           ###   ########.fr       */
+/*   Updated: 2025/02/24 15:09:21 by gcrisp           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "mlx_util.h"
 
-void	put_pixel(t_img *img, int x, int y, int col)
+void	put_pixel(t_img *img, size_t x, size_t y, int col)
 {
 	char	*dst;
